+++ src/lib/domain/repositories/settings_repository.dart
@@ -4,6 +4,7 @@ abstract class SettingsInfoRepository {
   Stream<List<String>> get libraryFoldersStream;
   ValueStream<String> get fileExtensionsStream;
   ValueStream<bool> get playAlbumsInOrderStream;
+  ValueStream<int> get listenedPercentageStream;
 }
 
 abstract class SettingsRepository extends SettingsInfoRepository {
@@ -11,4 +12,5 @@ abstract class SettingsRepository extends SettingsInfoRepository {
   Future<void> removeLibraryFolder(String? path);
   Future<void> setFileExtension(String extensions);
   Future<void> setPlayAlbumsInOrder(bool playInOrder);
+  Future<void> setListenedPercentage(int percentage);
 }