+++ src/lib/system/repositories/settings_repository_impl.dart
@@ -7,12 +7,16 @@ class SettingsRepositoryImpl implements SettingsRepository {
   SettingsRepositoryImpl(this._settingsDataSource) {
     _settingsDataSource.fileExtensionsStream.listen(_fileExtensionsSubject.add);
     _settingsDataSource.playAlbumsInOrderStream.listen(_playAlbumsInOrderSubject.add);
+    _settingsDataSource.listenedPercentageStream.listen(
+      (value) => _listenedPercentageSubject.add(value),
+    );
   }
 
   final SettingsDataSource _settingsDataSource;
 
   final BehaviorSubject<String> _fileExtensionsSubject = BehaviorSubject();
   final BehaviorSubject<bool> _playAlbumsInOrderSubject = BehaviorSubject();
+  final BehaviorSubject<int> _listenedPercentageSubject = BehaviorSubject();
 
   @override
   Stream<List<String>> get libraryFoldersStream => _settingsDataSource.libraryFoldersStream;
@@ -43,5 +47,13 @@ class SettingsRepositoryImpl implements SettingsRepository {
   @override
   Future<void> setPlayAlbumsInOrder(bool playInOrder) async {
     await _settingsDataSource.setPlayAlbumsInOrder(playInOrder);
+  }
+
+  @override
+  ValueStream<int> get listenedPercentageStream => _listenedPercentageSubject.stream;
+
+  @override
+  Future<void> setListenedPercentage(int percentage) async {
+    await _settingsDataSource.setListenedPercentage(percentage);
   }
 }